requirements.txt
absl-py==1.0.0
APScheduler==3.6.3
astroid==2.9.0
astunparse==1.6.3
backports.zoneinfo==0.2.1
black==21.12b0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cycler==0.11.0
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.13.3
importlib-metadata==4.8.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
MedPy==0.4.0
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.12.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.2.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
SimpleITK==2.1.1
six==1.16.0
sklearn==0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-addons==0.15.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.3
tornado==6.1
typeguard==2.13.3
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
